sih_backend/requirements.txt
fastapi==0.115.7
uvicorn==0.34.0
SQLAlchemy==2.0.37
psycopg2-binary==2.9.10
python-dotenv==1.0.1
pandas==2.2.3
sentence-transformers==5.1.0
numpy==1.26.4
pgvector
langchain==0.3.15
langchain-groq==0.3.2
pydantic==2.11.7
langchain-community==0.3.15
PyPDF2
requests
